+++ web/app/themes/rebe/templates/content-workshops.php
@@ -1,0 +1,195 @@
+<div class="o-hero c-hero--workshops">
+	<div class="u-centering-wrapper u-image-container" style="background-image: url(<?php  echo get_template_directory_uri() . '/dist/images/workshops--berlin--grace.jpg'; ?>);">
+		<div class="u-centered">
+			<div class="o-wrapper">
+				<h2 class="o-heading c-hero__title">
+					Bring <strong>global perspectives</strong> to conversations about
+					migration, identity and racial justice
+				</h2>
+				<a class="o-btn c-btn--large c-btn--blueberry _js-lightbox-open--book-a-workshop" href="https://goo.gl/fjKYa5" target="_blank">
+					Book a workshop
+				</a>
+			</div>
+		</div>
+	</div>
+</div>
+<!-- / Workshops page hero -->
+
+<div class="c-workshops-intro">
+	<div class="o-wrapper">
+
+		<div class="c-workshops-intro__about">
+			<p class="o-paragraph o-lead">
+				<strong>reimaginebelonging.org</strong> is a project of the Wings and Roots initiative,
+				which offers
+				workshops that examine the intersections
+				of migration, racism and belonging through international, historical,
+				and personal perspectives.
+			</p>
+			<p class="o-paragraph c-lead--small">
+				To date, we have conducted <strong>over 150 workshops</strong> and <strong>screenings</strong> in countries including the United States, Germany, Poland, the Netherlands, the UK and Turkey.
+			</p>
+			<p class="o-paragraph">
+				We aim to empower youth and adult participants to explore their
+				own stories, challenge racism and exclusionary debates, and collectively
+				contribute to building more inclusive societies.
+			</p>
+			<br>
+			<div class="u-video-container">
+				<iframe src="https://player.vimeo.com/video/142780820" frameborder="0" webkitallowfullscreen mozallowfullscreen allowfullscreen></iframe>
+			</div>
+		</div>
+
+		<!-- .c-workshops-page__about -->
+
+		<div class="c-workshops-intro__featured">
+
+			<h2 class="o-heading c-heading--page-subtitle c-workshops-intro__featured-title">
+				Featured workshops
+			</h2>
+			<div class="c-workshop-preview c-workshop-preview--featured">
+				<div class="c-workshop-preview__head">
+					<h2 class="o-heading c-workshop-preview__title">
+						<strong>Exploring Historical and Personal Narratives of Migration</strong> Workshop
+					</h2>
+				</div>
+				<div class="c-workshop-preview__info">
+					<p class="o-paragraph c-workshop-preview__description">
+						Participants explore how current immigration debates relate to historical events and their own stories.					</p>
+					<ul class="o-list c-workshop-preview-specs__list">
+						<li class="c-workshop-preview-specs__item"><strong>Target group:</strong> Youth, adults</li>
+						<li class="c-workshop-preview-specs__item"><strong>Group size:</strong> 8-30</li>
+						<li class="c-workshop-preview-specs__item"><strong>Duration:</strong> 3-6 hours</li>
+					</ul>
+				</div>
+				<a class="o-btn c-btn--medium c-btn--blueberry _js-lightbox-open--book-a-workshop" href="https://goo.gl/fjKYa5" target="_blank">
+					Book now
+				</a>
+			</div>
+
+			<div class="c-workshop-preview c-workshop-preview--featured">
+				<div class="c-workshop-preview__head">
+					<h2 class="o-heading c-workshop-preview__title">
+						<strong>Where are you 'from' from?</strong> Screening and Discussion
+					</h2>
+				</div>
+				<div class="c-workshop-preview__info">
+					<p class="o-paragraph c-workshop-preview__description">
+						Through the voices of ten young people living in Berlin and New York, this 22-minute film highlights the insight of children of immigrants in two societies struggling with migration and national identity.
+					</p>
+					<ul class="o-list c-workshop-preview-specs__list">
+						<li class="c-workshop-preview-specs__item"><strong>Target group:</strong> Youth, adults</li>
+						<li class="c-workshop-preview-specs__item"><strong>Group size:</strong> 5+</li>
+						<li class="c-workshop-preview-specs__item"><strong>Duration:</strong> 1-3 hours</li>
+					</ul>
+				</div>
+				<a class="o-btn c-btn--medium c-btn--blueberry _js-lightbox-open--book-a-workshop" href="https://goo.gl/fjKYa5" target="_blank">
+					Book now
+				</a>
+			</div>
+
+		</div><!-- .c-workshops-intro__featured -->
+
+	</div> <!-- o-wrapper -->
+</div> <!-- c-workshops-page__intro -->
+
+
+<!-- List of workshops -->
+
+<div class="c-workshop-list">
+	<div class="o-wrapper">
+
+		<h2 class="o-heading c-heading--page-subtitle c-workshop-list__title">
+			Workshops
+		</h2>
+
+		<div class="c-workshop-preview">
+			<div class="c-workshop-preview__head">
+				<h2 class="o-heading c-workshop-preview__title">
+					<strong>Exploring Historical and Personal Narratives of Migration in the Classroom</strong> Workshop
+				</h2>
+			</div>
+			<div class="c-workshop-preview__info">
+				<p class="o-paragraph c-workshop-preview__description">
+					Teachers practice hands-on activities to integrate migration histories and narratives into their educational work.
+				</p>
+				<ul class="o-list c-workshop-preview-specs__list">
+					<li class="c-workshop-preview-specs__item"><strong>Target group:</strong> Teachers, educators</li>
+					<li class="c-workshop-preview-specs__item"><strong>Group size:</strong> 6-30</li>
+					<li class="c-workshop-preview-specs__item"><strong>Duration:</strong> 6-12 hours</li>
+				</ul>
+			</div>
+			<a class="o-btn c-btn--medium c-btn--blueberry _js-lightbox-open--book-a-workshop" href="https://goo.gl/fjKYa5" target="_blank">
+				Book now
+			</a>
+		</div>
+
+		<div class="c-workshop-preview">
+			<div class="c-workshop-preview__head">
+				<h2 class="o-heading c-workshop-preview__title">
+					<strong>Addressing Bias-based Bullying in Schools</strong> Workshop
+				</h2>
+			</div>
+			<div class="c-workshop-preview__info">
+				<p class="o-paragraph c-workshop-preview__description">
+					Students learn strategies to identify and challenge bias-based bullying using our award-winning short film Article of Faith.
+				</p>
+				<ul class="o-list c-workshop-preview-specs__list">
+					<li class="c-workshop-preview-specs__item"><strong>Target group:</strong> Youth</li>
+					<li class="c-workshop-preview-specs__item"><strong>Group size:</strong> 10-30</li>
+					<li class="c-workshop-preview-specs__item"><strong>Duration:</strong> 1-6 hours</li>
+				</ul>
+			</div>
+			<a class="o-btn c-btn--medium c-btn--blueberry _js-lightbox-open--book-a-workshop" href="https://goo.gl/fjKYa5" target="_blank">
+				Book now
+			</a>
+		</div>
+
+		<div class="c-workshop-preview">
+			<div class="c-workshop-preview__head">
+				<h2 class="o-heading c-workshop-preview__title">
+					<strong>Migration, Media and Representation</strong> Youth Workshop
+				</h2>
+			</div>
+			<div class="c-workshop-preview__info">
+				<p class="o-paragraph c-workshop-preview__description">
+					These intensive workshops for youth combine film production with media analysis and biography work.
+				</p>
+				<ul class="o-list c-workshop-preview-specs__list">
+					<li class="c-workshop-preview-specs__item"><strong>Target group:</strong> Youth</li>
+					<li class="c-workshop-preview-specs__item"><strong>Group size:</strong> 8-20</li>
+					<li class="c-workshop-preview-specs__item"><strong>Duration:</strong> 3-10 days</li>
+				</ul>
+			</div>
+			<a class="o-btn c-btn--medium c-btn--blueberry _js-lightbox-open--book-a-workshop" href="https://goo.gl/fjKYa5" target="_blank">
+				Book now
+			</a>
+		</div>
+
+		<div class="c-workshop-preview">
+			<div class="c-workshop-preview__head">
+				<h2 class="o-heading c-workshop-preview__title">
+					<strong>Article of Faith</strong> Screening and Discussion
+				</h2>
+			</div>
+			<div class="c-workshop-preview__info">
+				<p class="o-paragraph c-workshop-preview__description">
+					We screen and discuss this award-winning 10-minute documentary about Sikh-American activist Sonny Singh and his efforts to combat bias-based school bullying.
+				</p>
+				<ul class="o-list c-workshop-preview-specs__list">
+					<li class="c-workshop-preview-specs__item"><strong>Target group:</strong> Youth, adults</li>
+					<li class="c-workshop-preview-specs__item"><strong>Group size:</strong> 5+</li>
+					<li class="c-workshop-preview-specs__item"><strong>Duration:</strong> 1-2 hours</li>
+				</ul>
+			</div>
+			<a class="o-btn c-btn--medium c-btn--blueberry _js-lightbox-open--book-a-workshop" href="https://goo.gl/fjKYa5" target="_blank">
+				Book now
+			</a>
+		</div>
+
+	</div><!-- .o-wrapper -->
+</div><!-- .c-workshop-list -->
+
+<!-- Book a Workshop Modal -->
+
+<?php get_template_part('templates/modal', 'book-a-workshop'); ?>