+++ web/app/themes/rebe/templates/footer.php
@@ -22,5 +22,4 @@
     </nav>
   </div>
 </footer>
-
 <?php get_template_part('templates/modal', 'donate'); ?>