+++ web/app/themes/rebe/templates/modal-book-a-workshop.php
@@ -1,0 +1,27 @@
+<div class="o-lightbox c-lightbox--book-a-workshop _js-lightbox--book-a-workshop">
+  <div class="o-lightbox__content-wrapper c-lightbox__content-wrapper--donate">
+
+    <div class="c-lightbox__nav c-lightbox__nav--close o-icon c-icon-close-medium-gray js-lightbox-close">
+      <?php the_field('close_button_label', 'option'); ?>
+    </div>
+
+    <h1 class="o-heading c-donate-modal__title">
+      Book A Workshop
+    </h1>
+    <p class="o-paragraph c-donate-modal__text">
+      Some text
+    </p>
+    <a class="o-btn c-btn--large c-btn--indigo" a href="#">
+      Book now
+    </a>
+    <div class="xxxxx">
+      <form>
+        First name:<br>
+        <input type="text" name="firstname"><br>
+        Last name:<br>
+        <input type="text" name="lastname">
+      </form>
+    </div>
+
+  </div>
+</div>